# openapi.yaml
openapi: 3.0.0
info:
  title: Test API
  version: 1.0.0
  description: API for managing test posts

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /api/test/{id}:
    get:
      summary: Retrieve a single user by ID
      description: |
        Fetches a single user from the system based on the provided `id` parameter.
        This endpoint returns user details, including ID, name, email, and role.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user to retrieve
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Successfully retrieved user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: "John Doe"
                  email:
                    type: string
                    format: email
                    example: "johndoe@example.com"
                  role:
                    type: string
                    enum: [admin, user, guest]
                    example: "user"
        "400":
          description: Invalid ID supplied (must be a number)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format. ID must be a number."
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User with ID 1 not found."
      security:
        - BearerAuth: []
      tags:
        - Users

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT